move oracle config for tokens to tokens config
