+++ config/tokens.ts
@@ -1,7 +1,32 @@
 import { ethers } from "ethers";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { getSyntheticTokenAddress } from "../utils/token";
+import { decimalToFloat } from "../utils/math";
 import { OracleProvider } from "./oracle";
+import { BigNumberish } from "ethers";
+import { TOKEN_ORACLE_TYPES } from "../utils/oracle";
+
+export type OracleProvider = "gmOracle" | "chainlinkDataStream" | "chainlinkPriceFeed";
+
+type OracleRealPriceFeed = {
+  address: string;
+  decimals: number;
+  heartbeatDuration: number;
+  stablePrice?: BigNumberish;
+  deploy?: never;
+  initPrice?: never;
+};
+
+type OracleTestPriceFeed = {
+  address?: never;
+  decimals: number;
+  heartbeatDuration: number;
+  stablePrice?: BigNumberish;
+  deploy: true;
+  initPrice: string;
+};
+
+type OraclePriceFeed = OracleRealPriceFeed | OracleTestPriceFeed;
 
 // synthetic token without corresponding token
 // address will be generated in runtime in hardhat.config.ts
@@ -18,6 +43,8 @@ type SyntheticTokenConfig = {
   dataStreamFeedDecimals?: number;
   oracleProvider?: OracleProvider;
   oracleTimestampAdjustment?: number;
+  priceFeed?: OraclePriceFeed;
+  oracleType?: string;
 };
 
 type RealTokenConfig = {
@@ -62,12 +89,22 @@ const config: {
       dataStreamFeedId: "0x000316d702a8e25e6b4ef4d449e3413dff067ee77dd366f0550251c07daf05ee",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x221912ce795669f628c51c69b7d0873eDA9C03bB",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     BTC: {
       synthetic: true,
       decimals: 8,
       dataStreamFeedId: "0x00039d9e45394f473ab1f050a1b963e6b05351e52d71e507509ada0c95ed75b8",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x6ce185860a4963106506C203335A2910413708e9",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     "WBTC.e": {
       address: "0x2f2a2543B76A4166549F7aaB2e75Bef0aefC5B0f",
@@ -75,6 +112,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x00039d9e45394f473ab1f050a1b963e6b05351e52d71e507509ada0c95ed75b8",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        // use the BTC price feed since the oracle would report the BTC price as well
+        address: "0x6ce185860a4963106506C203335A2910413708e9",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     tBTC: {
       address: "0x6c84a8f1c29108f47a79964b5fe888d4f4d0de40",
@@ -82,6 +125,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x00039d9e45394f473ab1f050a1b963e6b05351e52d71e507509ada0c95ed75b8",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        // use the BTC price feed since the oracle would report the BTC price as well
+        address: "0x6ce185860a4963106506C203335A2910413708e9",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     WETH: {
       address: "0x82aF49447D8a07e3bd95BD0d56f35241523fBab1",
@@ -90,6 +139,11 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x000362205e10b3a147d02792eccee483dca6c7b44ecce7012cb8c6e0b68b3ae9",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x639Fe6ab55C921f74e7fac1ee960C0B6293ba612",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     wstETH: {
       address: "0x5979D7b546E38E414F7E9822514be443A4800529",
@@ -106,6 +160,11 @@ const config: {
       dataStreamFeedId: "0x000335fd3f3ffa06cfd9297b97367f77145d7a5f132e84c736cc471dd98621fe",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x6970460aabF80C5BE983C6b74e5D06dEDCA95D4A",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     XRP: {
       synthetic: true,
@@ -120,6 +179,11 @@ const config: {
       dataStreamFeedId: "0x000356ca64d3b32135e17dc0dc721a645bf50d0303be8ceb2cdca0a50bab8fdc",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x9A7FB1b3950837a8D9b40517626E11D4127C098C",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     EIGEN: {
       address: "0x606c3e5075e5555e79aa15f1e9facb776f96c248",
@@ -143,6 +207,11 @@ const config: {
       dataStreamFeedId: "0x0003b778d3f6b2ac4991302b89cb313f99a42467d6c9c5f96f57c29c0d2bc24f",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x24ceA4b8ce57cdA5058b924B9B9987992450590c",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     STX: {
       synthetic: true,
@@ -172,6 +241,11 @@ const config: {
       dataStreamFeedId: "0x000367a3674cdd7cc83dbbd7d19f3768b9d1329586e82e32a1bf388fc5ffd0eb",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x9C917083fDb403ab5ADbEC26Ee294f6EcAda2720",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     LINK: {
       address: "0xf97f4df75117a78c1A5a0DBb814Af92458539FB4",
@@ -180,6 +254,11 @@ const config: {
       dataStreamFeedId: "0x00036d7a1251e3f67d6658466b5e9e7fe8418af7feac9567ff322bff95cc2401",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x86E53CF1B870786351Da77A57575e79CB55812CB",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     ARB: {
       address: "0x912CE59144191C1204E64559FE8253a0e49E6548",
@@ -188,6 +267,11 @@ const config: {
       dataStreamFeedId: "0x00030ab7d02fbba9c6304f98824524407b1f494741174320cfd17a2c22eec1de",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xb2A824043730FE05F3DA2efaFa1CBbe83fa548D6",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     ATOM: {
       synthetic: true,
@@ -195,6 +279,11 @@ const config: {
       dataStreamFeedId: "0x0003c93d9477344d0f2516c4605008399e6750d492a60ab85a9ecb2e441e87b3",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xCDA67618e51762235eacA373894F0C79256768fa",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     NEAR: {
       synthetic: true,
@@ -202,6 +291,11 @@ const config: {
       dataStreamFeedId: "0x00036e9386eda6b177c6f7e9d493e60ae9ebaeb732a271b880b4d6a131d6b3f5",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xBF5C3fB2633e924598A46B9D07a174a9DBcF57C0",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     AAVE: {
       address: "0xba5ddd1f9d7f570dc94a51479a000e3bce967196",
@@ -210,6 +304,11 @@ const config: {
       dataStreamFeedId: "0x0003481a2f7fe21c01d427f39035541d2b7a53db9c76234dc36082e6ad6db7f5",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xaD1d5344AaDE45F43E596773Bcc4c423EAbdD034",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     AVAX: {
       address: "0x565609fAF65B92F7be02468acF86f8979423e514",
@@ -218,6 +317,11 @@ const config: {
       dataStreamFeedId: "0x0003b1cb55b8a00a18111ff745b56d70f04c2d41e03fc7fd8d3d9b09f142aad9",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x8bf61728eeDCE2F32c456454d87B5d6eD6150208",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     OP: {
       address: "0xaC800FD6159c2a2CB8fC31EF74621eB430287a5A",
@@ -226,6 +330,11 @@ const config: {
       dataStreamFeedId: "0x0003ba5e477371c9cd5446281354a91e66c233dd50e26efe6abbfdc310b92dab",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x205aaD468a11fd5D34fA7211bC6Bad5b3deB9b98",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     ORDI: {
       synthetic: true,
@@ -241,6 +350,11 @@ const config: {
       dataStreamFeedId: "0x0003169a4ebb9178e5ec6281913d1a8a4f676f414c94b60a4cb2e432f9081c60",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xDB98056FecFff59D032aB628337A4887110df3dB",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     PEPE: {
       address: "0x25d887Ce7a35172C62FeBFD67a1856F20FaEbB00",
@@ -249,6 +363,11 @@ const config: {
       dataStreamFeedId: "0x00032fd010bde9fb3b8b53c126bd8a0bd2c3b3fbaceb33605cd46a7e461e4da8",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x02DEd5a7EDDA750E3Eb240b54437a54d57b74dBE",
+        decimals: 18,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     WIF: {
       address: "0xA1b91fe9FD52141Ff8cac388Ce3F10BFDc1dE79d",
@@ -257,6 +376,11 @@ const config: {
       dataStreamFeedId: "0x0003d5a1e39f957e312e307d535c8c28315172442ae0c39d9488e908c3762c85",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xF7Ee427318d2Bd0EEd3c63382D0d52Ad8A68f90D",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     USDC: {
       address: "0xaf88d065e77c8cC2239327C5EDb3A432268e5831",
@@ -264,6 +388,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x00038f83323b6b08116d1614cf33a9bd71ab5e0abf0c9f1b783a74a43e7bd992",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x50834F3163758fcC1Df9973b6e91f0F0F0434aD3",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     "USDC.e": {
       address: "0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8",
@@ -271,6 +401,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x00038f83323b6b08116d1614cf33a9bd71ab5e0abf0c9f1b783a74a43e7bd992",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x50834F3163758fcC1Df9973b6e91f0F0F0434aD3",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     USDT: {
       address: "0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9",
@@ -279,6 +415,12 @@ const config: {
       dataStreamFeedId: "0x0003a910a43485e0685ff5d6d366541f5c21150f0634c5b14254392d1a1c06db",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x3f3f5dF88dC9F13eac63DF89EC16ef6e7E25DdE7",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     DAI: {
       address: "0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1",
@@ -287,6 +429,12 @@ const config: {
       dataStreamFeedId: "0x0003571a596aab1cafe1a883795dab90cd8cca48e25f53ab3665aa67d008f386",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xc5C8E77B397E531B8EC06BFb0048328B30E9eCfB",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     USDe: {
       address: "0x5d3a1Ff2b6BAb83b63cd9AD0787074081a52ef34",
@@ -295,6 +443,12 @@ const config: {
       dataStreamFeedId: "0x000379477abdca006db92300211242ba44479369a15be0ad8d59cabe00e63074",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x88AC7Bca36567525A866138F03a6F6844868E0Bc",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
   },
   avalanche: {
@@ -304,6 +458,11 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x00039d9e45394f473ab1f050a1b963e6b05351e52d71e507509ada0c95ed75b8",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x2779D32d5166BAaa2B2b658333bA7e6Ec0C65743",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     "WETH.e": {
       address: "0x49D5c2BdFfac6CE2BFdB6640F4F80f226bc10bAB",
@@ -311,6 +470,11 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x000362205e10b3a147d02792eccee483dca6c7b44ecce7012cb8c6e0b68b3ae9",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x976B3D034E162d8bD72D6b9C989d545b839003b0",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     XRP: {
       synthetic: true,
@@ -349,6 +513,11 @@ const config: {
       dataStreamFeedId: "0x0003b1cb55b8a00a18111ff745b56d70f04c2d41e03fc7fd8d3d9b09f142aad9",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x0A77230d17318075983913bC2145DB16C7366156",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+      },
     },
     USDC: {
       address: "0xb97ef9ef8734c71904d8002f8b6bc66dd9c48a6e",
@@ -356,6 +525,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x00038f83323b6b08116d1614cf33a9bd71ab5e0abf0c9f1b783a74a43e7bd992",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0xF096872672F44d6EBA71458D74fe67F9a77a23B9",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     "USDC.e": {
       address: "0xA7D7079b0FEaD91F3e65f86E8915Cb59c1a4C664",
@@ -363,6 +538,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x00038f83323b6b08116d1614cf33a9bd71ab5e0abf0c9f1b783a74a43e7bd992",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0xF096872672F44d6EBA71458D74fe67F9a77a23B9",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     USDT: {
       address: "0x9702230A8Ea53601f5cD2dc00fDBc13d4dF4A8c7",
@@ -371,6 +552,12 @@ const config: {
       dataStreamFeedId: "0x0003a910a43485e0685ff5d6d366541f5c21150f0634c5b14254392d1a1c06db",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xEBE676ee90Fe1112671f19b6B7459bC678B67e8a",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     "USDT.e": {
       address: "0xc7198437980c041c805A1EDcbA50c1Ce5db95118",
@@ -379,6 +566,12 @@ const config: {
       dataStreamFeedId: "0x0003a910a43485e0685ff5d6d366541f5c21150f0634c5b14254392d1a1c06db",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0xEBE676ee90Fe1112671f19b6B7459bC678B67e8a",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     "DAI.e": {
       address: "0xd586E7F844cEa2F87f50152665BCbc2C279D8d70",
@@ -387,6 +580,12 @@ const config: {
       dataStreamFeedId: "0x0003571a596aab1cafe1a883795dab90cd8cca48e25f53ab3665aa67d008f386",
       dataStreamFeedDecimals: 18,
       oracleTimestampAdjustment: 1,
+      priceFeed: {
+        address: "0x51D7180edA2260cc4F6e4EebB82FEF5c3c2B8300",
+        decimals: 8,
+        heartbeatDuration: (24 + 1) * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
   },
   arbitrumSepolia: {
@@ -404,6 +603,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x0001ef0379c835e64d03082e561403cc14b2779b525d93149b25df0ee3ef9456",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x56a43EB56Da12C0dc1D972ACb089c06a5dEF8e69",
+        decimals: 8,
+        heartbeatDuration: 3 * 24 * 60 * 60,
+        stablePrice: decimalToFloat(44000),
+      },
     },
     USDC: {
       address: "0x3321Fd36aEaB0d5CdfD26f4A3A93E2D2aAcCB99f",
@@ -411,6 +616,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x0001829d7d4c7c5badcd54d2126e621c51eabf32393ffab969e311b18ed80138",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x0153002d20B96532C639313c2d54c3dA09109309",
+        decimals: 8,
+        heartbeatDuration: 3 * 24 * 60 * 60,
+        stablePrice: decimalToFloat(1),
+      },
     },
     GLV_WETH_USDC: {
       address: "0x528A5bac7E746C9A509A1f4F6dF58A03d44279F9",
@@ -446,6 +657,11 @@ const config: {
       address: "0x04FC936a15352a1b15b3B9c56EA002051e3DB3e5",
       dataStreamFeedId: "0x555344432d5553442d415242495452554d2d544553544e455400000000000000",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x1692Bdd32F31b831caAc1b0c9fAF68613682813b",
+        decimals: 8,
+        heartbeatDuration: 3 * 24 * 60 * 60,
+      },
     },
     SOL: {
       synthetic: true,
@@ -457,6 +673,11 @@ const config: {
       address: "0xBFcBcdCbcc1b765843dCe4DF044B92FE68182a62",
       dataStreamFeedId: "0x12be1859ee43f46bab53750915f20855f54e891f88ddd524f26a72d6f4deed1d",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x0a023a3423D9b27A0BE48c768CCF2dD7877fEf5E",
+        decimals: 8,
+        heartbeatDuration: 3 * 24 * 60 * 60,
+      },
     },
     DAI: {
       decimals: 18,
@@ -464,6 +685,11 @@ const config: {
       address: "0x7b7c6c49fA99b37270077FBFA398748c27046984",
       dataStreamFeedId: "0xbf1febc8c335cb236c1995c1007a928a3f7ae8307a1a20cb31334e6d316c62d1",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        address: "0x103b53E977DA6E4Fa92f76369c8b7e20E7fb7fe1",
+        decimals: 8,
+        heartbeatDuration: 3 * 24 * 60 * 60,
+      },
     },
     TEST: {
       synthetic: true,
@@ -542,6 +768,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x0003dc85e8b01946bf9dfd8b0db860129181eb6105a8c8981d9f28e00b6f60d9",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        // this is USDT price feed, there is no USDC feed on Avalanche Fuji
+        address: "0x7898AcCC83587C3C55116c5230C17a6Cd9C71bad",
+        decimals: 8,
+        heartbeatDuration: 3 * 24 * 60 * 60,
+      },
     },
     USDT: {
       decimals: 6,
@@ -549,6 +781,12 @@ const config: {
       transferGasLimit: 200 * 1000,
       dataStreamFeedId: "0x00032874077216155926e26c159c1c20a572921371d9de605fe9633e48d136f9",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        // this is USDT price feed, there is no USDC feed on Avalanche Fuji
+        address: "0x7898AcCC83587C3C55116c5230C17a6Cd9C71bad",
+        decimals: 8,
+        heartbeatDuration: 3 * 24 * 60 * 60,
+      },
     },
     DAI: {
       decimals: 6,
@@ -556,6 +794,12 @@ const config: {
       address: "0x51290cb93bE5062A6497f16D9cd3376Adf54F920",
       dataStreamFeedId: "0x0003649272a19e143a7f4c2d98905b413e98dce81fb09287dcf4c513cba5cc72",
       dataStreamFeedDecimals: 18,
+      priceFeed: {
+        // this is USDT price feed, there is no USDC feed on Avalanche Fuji
+        address: "0x7898AcCC83587C3C55116c5230C17a6Cd9C71bad",
+        decimals: 8,
+        heartbeatDuration: 3 * 24 * 60 * 60,
+      },
     },
     WETH: {
       address: "0x82F0b3695Ed2324e55bbD9A9554cB4192EC3a514",
@@ -620,6 +864,12 @@ const config: {
       decimals: 18,
       transferGasLimit: 200 * 1000,
       deploy: true,
+      priceFeed: {
+        decimals: 8,
+        heartbeatDuration: 24 * 60 * 60,
+        deploy: true,
+        initPrice: "500000000000",
+      },
     },
     WBTC: {
       decimals: 8,
@@ -630,11 +880,23 @@ const config: {
       decimals: 6,
       transferGasLimit: 200 * 1000,
       deploy: true,
+      priceFeed: {
+        decimals: 8,
+        heartbeatDuration: 24 * 60 * 60,
+        deploy: true,
+        initPrice: "100000000",
+      },
     },
     USDT: {
       decimals: 6,
       transferGasLimit: 200 * 1000,
       deploy: true,
+      priceFeed: {
+        decimals: 8,
+        heartbeatDuration: 24 * 60 * 60,
+        deploy: true,
+        initPrice: "100000000",
+      },
     },
     SOL: {
       synthetic: true,
@@ -683,6 +945,10 @@ export default async function (hre: HardhatRuntimeEnvironment): Promise<TokensCo
     }
     if (!hre.network.live) {
       (token as any).deploy = true;
+    }
+
+    if (token.oracleType === undefined) {
+      token.oracleType = TOKEN_ORACLE_TYPES.DEFAULT;
     }
   }
